refactor: move key backlight to the module
